+++ apps/admin/src/app/_api/types.ts
@@ -1,0 +1,8 @@
+import type { IconType } from '@repo/ui/components/Icon'
+
+export type Request = {
+  placeId: string
+  placeName: string
+  icon: IconType
+  requestDate: string
+}
