+++ apps/admin/src/app/request/[id]/_api/types.ts
@@ -17,13 +17,6 @@ type Tag = {
   iconKey: IconType
 }
 
-export type Request = {
-  placeId: string
-  placeName: string
-  icon: IconType
-  requestDate: string
-}
-
 export type RequestDetail = {
   placeId: string
   placeName: string
